+++ f3d/src/main/java/app/f3d/F3D/android/MainActivity.kt
@@ -1,35 +1,21 @@
 package app.f3d.F3D.android
 
 import android.content.Intent
-import android.graphics.Color
-import android.graphics.drawable.GradientDrawable
 import android.net.Uri
 import android.os.Bundle
-import android.util.TypedValue
-import android.view.Gravity
 import android.view.View
 import android.view.ViewGroup
 import android.view.ViewOutlineProvider
 import android.widget.ImageButton
-import android.widget.ImageView
-import android.widget.LinearLayout
-import android.widget.LinearLayout.LayoutParams
-import android.widget.TextView
 import androidx.activity.result.ActivityResultLauncher
-import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
-import androidx.browser.customtabs.CustomTabColorSchemeParams
-import androidx.browser.customtabs.CustomTabsIntent
 import androidx.coordinatorlayout.widget.CoordinatorLayout
-import androidx.core.graphics.ColorUtils
-import androidx.core.net.toUri
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowCompat
 import androidx.core.view.WindowInsetsCompat
 import app.f3d.F3D.android.Utils.ConsoleLog
 import app.f3d.F3D.android.Utils.FileInteractionContract
 import com.google.android.material.bottomappbar.BottomAppBar
-import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.google.android.material.floatingactionbutton.FloatingActionButton
 import com.google.android.material.shape.MaterialShapeDrawable
 
@@ -75,7 +61,7 @@ class MainActivity : AppCompatActivity() {
         }
 
         findViewById<ImageButton>(R.id.supportButton).setOnClickListener { _: View? ->
-            showSupportDialog()
+            SupportDialog(this).show()
         }
 
         handleSelectedFileAppNotOpen()
@@ -167,115 +153,6 @@ class MainActivity : AppCompatActivity() {
             }
             show()
         }
-    }
-
-    private fun dp(value: Int): Int =
-        (value * resources.displayMetrics.density).toInt()
-
-    private fun showSupportDialog() {
-        val primaryColor = getColor(R.color.yellow)
-        val textColor = getColor(R.color.white)
-        val backgroundColor = getColor(R.color.panel_surface)
-
-        // Heart icon
-        val icon = ImageView(this).apply {
-            setImageResource(R.drawable.ic_baseline_support_24)
-            imageTintList = android.content.res.ColorStateList.valueOf(getColor(R.color.red))
-            layoutParams = LayoutParams(dp(56), dp(56)).apply {
-                gravity = Gravity.CENTER
-                bottomMargin = dp(16)
-            }
-        }
-
-        // Title
-        val title = TextView(this).apply {
-            text = context.getString(R.string.support_f3d)
-            setTextColor(textColor)
-            textSize = 22f
-            setTypeface(typeface, android.graphics.Typeface.BOLD)
-            gravity = Gravity.CENTER
-            layoutParams = LayoutParams(
-                ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.WRAP_CONTENT
-            ).apply { bottomMargin = dp(12) }
-        }
-
-        // Description
-        val desc = TextView(this).apply {
-            text = context.getString(R.string.support_f3d_description)
-            setTextColor(ColorUtils.setAlphaComponent(textColor, 192))
-            textSize = 15f
-            gravity = Gravity.CENTER
-            layoutParams = LayoutParams(
-                ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.WRAP_CONTENT
-            ).apply { bottomMargin = dp(24) }
-        }
-
-        // Donate button
-        val donateBtn = TextView(this).apply {
-            text = context.getString(R.string.donate)
-            setTextColor(getColor(R.color.black))
-            textSize = 16f
-            setTypeface(typeface, android.graphics.Typeface.BOLD)
-            gravity = Gravity.CENTER
-            setPadding(dp(24), dp(14), dp(24), dp(14))
-            background = GradientDrawable().apply {
-                cornerRadius = dp(50).toFloat()
-                setColor(primaryColor)
-            }
-            layoutParams = LayoutParams(
-                ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.WRAP_CONTENT
-            ).apply { bottomMargin = dp(10) }
-        }
-
-        // Dismiss link
-        val dismissBtn = TextView(this).apply {
-            text = context.getString(R.string.maybe_later)
-            setTextColor(ColorUtils.setAlphaComponent(textColor, 128))
-            textSize = 14f
-            gravity = Gravity.CENTER
-            setPadding(0, dp(6), 0, dp(6))
-        }
-
-        val content = LinearLayout(this).apply {
-            orientation = LinearLayout.VERTICAL
-            background = GradientDrawable().apply {
-                cornerRadius = dp(24).toFloat()
-                setColor(backgroundColor)
-            }
-            setPadding(dp(28), dp(32), dp(28), dp(20))
-            addView(icon)
-            addView(title)
-            addView(desc)
-            addView(donateBtn)
-            addView(dismissBtn)
-        }
-
-        val dialog = MaterialAlertDialogBuilder(this)
-            .setBackground(GradientDrawable().apply { setColor(Color.TRANSPARENT) })
-            .setView(content)
-            .create()
-        dialog.window?.setDimAmount(0.5f)
-
-        donateBtn.setOnClickListener {
-            dialog.dismiss()
-            showDonateWebView()
-        }
-        dismissBtn.setOnClickListener { dialog.dismiss() }
-
-        dialog.show()
-    }
-
-    private fun showDonateWebView() {
-        CustomTabsIntent.Builder()
-            .setColorScheme(CustomTabsIntent.COLOR_SCHEME_DARK)
-            .setDefaultColorSchemeParams(
-                CustomTabColorSchemeParams.Builder()
-                    .setToolbarColor(getColor(R.color.panel_surface))
-                    .build()
-            )
-            .setShowTitle(false)
-            .build()
-            .launchUrl(this, "https://donate.stripe.com/4gM00j7tO1w4eAD5J0cs800".toUri())
     }
 
     private fun setChromeVisible(visible: Boolean) {
